requirements.txt
Django==1.3.1
distribute==0.6.24
django-social-auth==0.6.1
facebook-python-sdk==0.1
httplib2==0.7.2
oauth2==1.5.211
python-openid==2.2.5
wsgiref==0.1.2
